Add references to RFCs defining good documentation practice
In these RFCs, great details and example are given about the right way to do
documentation. I’m not sure if this is worth being detailed on
cheats.rs, but a link is a way as good as another to find these
documents.

In particular, these RFC are way more detailed than the Reference book.

See
https://github.com/rust-lang/rust/blob/master/CONTRIBUTING.md#writing-documentation
for a proof that RFC1574 is used for std.